• Separates TableBody
