+++ src/Components/SummaryTable/SummaryTable.tsx
@@ -1,20 +1,19 @@
 import React from 'react';
 import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
 import TablePagination from '@material-ui/core/TablePagination';
-import TableRow from '@material-ui/core/TableRow';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
 import Paper from '@material-ui/core/Paper';
-import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Switch from '@material-ui/core/Switch';
 import SummaryTableToolBar from './SummaryTableToolBar/SummaryTableToolBar';
 import SummaryTableHeader from './SummaryTableHeader/SummaryTableHeader';
+import SummaryTableBody from './SummaryTableBody/SummaryTableBody';
 
-interface Data {
+interface Props {
+  
+}
+
+export type SummaryTableRowData = {
   amount: number;
   name: string;
 }
@@ -22,7 +21,7 @@ interface Data {
 function createData(
   name: string,
   amount: number
-): Data {
+): SummaryTableRowData {
   return { name, amount };
 }
 
@@ -42,34 +41,7 @@ const rows = [
   createData('Oreo', 437),
 ];
 
-function desc<T>(a: T, b: T, orderBy: keyof T) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function stableSort<T>(array: T[], cmp: (a: T, b: T) => number) {
-  const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-  stabilizedThis.sort((a, b) => {
-    const order = cmp(a[0], b[0]);
-    if (order !== 0) return order;
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map(el => el[0]);
-}
-
-type Order = 'asc' | 'desc';
-
-function getSorting<K extends keyof any>(
-  order: Order,
-  orderBy: K,
-): (a: { [key in K]: number | string }, b: { [key in K]: number | string }) => number {
-  return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
-}
+export type SummaryTableSortingOrder = 'asc' | 'desc';
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -90,16 +62,16 @@ const useStyles = makeStyles((theme: Theme) =>
   }),
 );
 
-export default function SummaryTable() {
+export default function SummaryTable(props: Props) {
   const classes = useStyles();
-  const [order, setOrder] = React.useState<Order>('asc');
-  const [orderBy, setOrderBy] = React.useState<keyof Data>('amount');
+  const [order, setOrder] = React.useState<SummaryTableSortingOrder>('asc');
+  const [orderBy, setOrderBy] = React.useState<keyof SummaryTableRowData>('amount');
   const [selected, setSelected] = React.useState<string[]>([]);
   const [page, setPage] = React.useState(0);
   const [dense, setDense] = React.useState(false);
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
-  function handleRequestSort(event: React.MouseEvent<unknown>, property: keyof Data) {
+  function handleRequestSort(event: React.MouseEvent<unknown>, property: keyof SummaryTableRowData) {
     const isDesc = orderBy === property && order === 'desc';
     setOrder(isDesc ? 'asc' : 'desc');
     setOrderBy(property);
@@ -114,7 +86,7 @@ export default function SummaryTable() {
     setSelected([]);
   }
 
-  function handleClick(event: React.MouseEvent<unknown>, name: string) {
+  function handleRowSelected(event: React.MouseEvent<unknown>, name: string) {
     const selectedIndex = selected.indexOf(name);
     let newSelected: string[] = [];
 
@@ -147,10 +119,6 @@ export default function SummaryTable() {
     setDense(event.target.checked);
   }
 
-  const isSelected = (name: string) => selected.indexOf(name) !== -1;
-
-  const emptyRows = rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage);
-
   return (
     <div className={classes.root}>
       <Paper className={classes.paper}>
@@ -169,42 +137,15 @@ export default function SummaryTable() {
               onRequestSort={handleRequestSort}
               rowCount={rows.length}
             />
-            <TableBody>
-              {stableSort(rows, getSorting(order, orderBy))
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .map((row, index) => {
-                  const isItemSelected = isSelected(row.name);
-                  const labelId = `enhanced-table-checkbox-${index}`;
-
-                  return (
-                    <TableRow
-                      hover
-                      role="checkbox"
-                      aria-checked={isItemSelected}
-                      tabIndex={-1}
-                      key={row.name}
-                      selected={isItemSelected}
-                      >
-                      <TableCell padding="checkbox">
-                        <Checkbox
-                          checked={isItemSelected}
-                          inputProps={{ 'aria-labelledby': labelId }}
-                          onClick={event => handleClick(event, row.name)}
-                        />
-                      </TableCell>
-                      <TableCell component="th" id={labelId} scope="row" padding="none">
-                        {row.name}
-                      </TableCell>
-                      <TableCell align="right">{row.amount}</TableCell>
-                    </TableRow>
-                  );
-                })}
-              {emptyRows > 0 && (
-                <TableRow style={{ height: 49 * emptyRows }}>
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
+            <SummaryTableBody 
+              rows={rows}
+              rowsPerPage={rowsPerPage}
+              selected={selected}
+              order={order}
+              orderBy={orderBy}
+              page={page}
+              handleSelectClick={handleRowSelected}
+            />
           </Table>
         </div>
         <TablePagination
